test case - inputs folder - 5

+++ test_cases/cli/inputs_folder/scripts/inputs_folder.sh
@@ -12,9 +12,16 @@ echo "Test case started: inputs_folder"
 echo "Copying input files ${INPUTS_SRC_DIR} -> ${INPUTS_TARGET_DIR}"
 mkdir -p ${INPUTS_TARGET_DIR}
 cp -avr ${INPUTS_TARGET_DIR} ${INPUTS_SRC_DIR}
+
 # run flow
 echo "Running test flow"
 ${CLI_EXECUTABLE} run --f ${TEST_FLOW}
+CLI_EXIT_CODE=$?
+if [ "${CLI_EXIT_CODE}" !=  "0" ]
+then
+  exit ${CLI_EXIT_CODE}
+fi
+
 # delete input files
 echo "Removing inputs folder: ${INPUTS_TARGET_DIR}"
 rm -r ${INPUTS_TARGET_DIR}
